- customise client appointment details screen  - customise work estimate in edit form for rent cars  - adding "complete button" in consultant appointment details screen

+++ lib/modules/work-estimate-form/screens/work-estimate-add-issue-modal.dart
@@ -236,7 +236,7 @@ class WorkEstimateAddIssueModalState extends State<WorkEstimateAddIssueModal> {
                               .toString()),
                       onChanged: (newQuantity) {
                         provider.estimatorFormState['quantity'] =
-                            int.parse(newQuantity);
+                            double.parse(newQuantity);
                       },
                       validator: (value) {
                         if (value == null) {
